# internal/apidocs/gen/swagger.json
{
    "components": {
        "schemas": {
...
            "server.machineNotesRequest": {
                "properties": {
                    "notes": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "server.machineNotesResponse": {
                "properties": {
                    "machine_name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "server.machineTagsRequest": {
                "properties": {
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "server.machineTagsResponse": {
                "properties": {
                    "machine_name": {
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "The swag-generated half of the Agent API v1 document. Serve-time authority lives in internal/apidocs: the hand-maintained fragment (openapi.json) owns info/tags/components/security and every not-yet-migrated path, and this generated document's paths merge into it.",
        "title": "Agent API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/machines/{machineName}/notes": {
            "get": {
                "description": "Minimum role: viewer.",
                "parameters": [
                    {
                        "description": "Machine name",
                        "in": "path",
                        "name": "machineName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.machineNotesResponse"
                                }
                            }
                        },
                        "description": "Notes"
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                },
                "summary": "Machine notes",
                "tags": [
                    "Machine Management"
                ]
            },
            "put": {
                "description": "Minimum role: operator. Registry-only, no task.",
                "parameters": [
                    {
                        "description": "Machine name",
                        "in": "path",
                        "name": "machineName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.machineNotesRequest",
                                        "summary": "request",
                                        "description": "Notes to set"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Notes to set",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Notes updated"
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                },
                "summary": "Update machine notes",
                "tags": [
                    "Machine Management"
                ]
            }
        },
        "/machines/{machineName}/tags": {
            "get": {
                "description": "Minimum role: viewer.",
                "parameters": [
                    {
                        "description": "Machine name",
                        "in": "path",
                        "name": "machineName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.machineTagsResponse"
                                }
                            }
                        },
                        "description": "Tags"
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                },
                "summary": "Machine tags",
                "tags": [
                    "Machine Management"
                ]
            },
            "put": {
                "description": "Minimum role: operator. Registry-only, no task.",
                "parameters": [
                    {
                        "description": "Machine name",
                        "in": "path",
                        "name": "machineName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.machineTagsRequest",
                                        "summary": "request",
                                        "description": "Tags to set"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Tags to set",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Tags updated"
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                },
                "summary": "Update machine tags",
                "tags": [
                    "Machine Management"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}